Move console into main server library
Since we're going to dynamically load only the server *functions*, it
makes the most sense to have the socket and session handling within
the main server library.  The server startup/cleanup portion becomes
massively simpler with this move.

Since we no longer have the console create factory, we can add a
simple no-argument start method to the Console itself.  We can also
dump the create/delete factory test.

There are a lot of scaffolding items around making this sort of move -
Makefiles, autoconf files, include paths, defines.  It's always
interesting to see just what has grown up around a given feature.

Re:  issue #263
