+++ configure.ac
@@ -513,7 +513,7 @@ my_bool some_var;
       [AC_MSG_ERROR([can not build without dlopen.])])
   LIBS="$DLOPEN_LIBS_SAVE"
 
-  SERVER_LDLIBS="$SERVER_LDLIBS $PROFILE_LDLIBS"
+  SERVER_LDLIBS="$SERVER_LDLIBS $PROFILE_LDLIBS $WRAP_LDLIBS"
 
   AC_MSG_NOTICE([using SERVER_LDLIBS=$SERVER_LDLIBS])
